requirements.txt
discord.py
aioconsole
quart
python-dotenv